+++ changelog.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Usability: Shortcuts for eg. Edit tool (**SPACE** select only one, **E** start editing, **D** delete feature)
   [docs](https://docs.webgiscloud.com/de/webgis/apps/viewer/customjs/usability.html#tastatur-shortcuts)
 
+- Usability: Tool Priority (sort tools inside containers)
+  [discussion #279](https://github.com/e-netze/webgis-community/discussions/279)
+  [docs](https://docs.webgiscloud.com/de/webgis/apps/viewer/customjs/usability.html#toolbox)
+
 - Usability: Edit Mask Combo/Select-Boxes - use ``select2``, for combos with 
   many items (combination ``domain``/``autocomplete``)